requirements.txt
uvicorn~=0.21.0
python-decouple
redis
starlette~=0.26.0.post1
fastapi~=0.94.0
passlib~=1.7.4
jwt~=1.3.1
pytz~=2022.7.1
requests~=2.28.2
mysql-connector-python~=8.0.31
itsdangerous~=2.1.2
email-validator
python-multipart~=0.0.5
prisma~=0.8.2
pydantic~=1.10.4
dnspython~=2.2.1
idna~=3.4
httpx~=0.23.3
cryptography~=39.0.0
future~=0.18.3
h11~=0.14.0
dataclasses~=0.6
pip~=22.3.1
attrs~=22.2.0
wheel~=0.38.4
docutils~=0.19
Jinja2~=3.1.2
keyring~=23.13.1
anyio~=3.6.2
sniffio~=1.3.0
click~=8.1.3
httpcore
Pygments~=2.14.0
rfc3986~=1.5.0
certifi~=2022.12.7
MarkupSafe~=2.1.2
protobuf~=3.20.1
tomlkit~=0.11.6
websockets
six~=1.16.0
urllib3~=1.26.14
pycparser~=2.21
setuptools~=65.5.0
packaging~=23.0
pyparsing~=3.0.9
zipp~=3.11.0
ruff